create endpoint get tour guide profile hiking point and api to toggle the...

See merge request enigma-camp/malang-branch/torvalds-batch-4/team/final-projects/team-4/giri-guide-backend!17

+++ src/main/java/com/abdav/giri_guide/model/response/TourGuideProfileResponse.java
@@ -15,10 +15,10 @@ public record TourGuideProfileResponse(
         String description,
         String address,
         Integer maxHiker,
-        Double price,
-        Double additionalPrice,
+        Long price,
+        Long additionalPrice,
         Integer totalPorter,
-        Double pricePorter,
+        Long pricePorter,
         String image
 
 ) {
